+++ src/degiro_portfolio/__init__.py
@@ -1,2 +1,2 @@
 """Stock price visualization application."""
-__version__ = "0.5.3"
+__version__ = "0.5.4"